refactor: move shooter and hitbox to same folder
