+++ game/src/game/GameApp.cpp
@@ -12,7 +12,6 @@
 #include "engine/input/InputService.h"
 #include "engine/physics/BoxRigidBody.h"
 #include "engine/physics/BoxTrigger.h"
-#include "engine/physics/Hitbox.h"
 #include "engine/physics/MeshStaticBody.h"
 #include "engine/physics/PhysicsService.h"
 #include "engine/physics/PlaneStaticBody.h"
@@ -32,11 +31,12 @@
 #include "game/components/Pickups/Powerups/EveryoneSlowerPickup.h"
 #include "game/components/Pickups/Powerups/IncreaseAimBoxPickup.h"
 #include "game/components/Pickups/Powerups/KillAbilitiesPickup.h"
-#include "game/components/Shooter.h"
 #include "game/components/VehicleComponent.h"
 #include "game/components/audio/AudioEmitter.h"
 #include "game/components/audio/AudioListener.h"
 #include "game/components/race/Checkpoint.h"
+#include "game/components/shooting/Hitbox.h"
+#include "game/components/shooting/Shooter.h"
 #include "game/components/state/PlayerState.h"
 #include "game/components/ui/PlayerHud.h"
 #include "game/services/GameStateService.h"
