+++ game/src/game/components/Controllers/PlayerController.cpp
@@ -10,7 +10,7 @@
 #include "engine/render/RenderService.h"
 #include "engine/scene/Entity.h"
 #include "engine/scene/Transform.h"
-#include "game/components/Shooter.h"
+#include "game/components/shooting/Shooter.h"
 #include "game/components/VehicleComponent.h"
 #include "game/components/state/PlayerState.h"
 #include "game/services/GameStateService.h"
