+++ game/src/game/services/GameStateService.cpp
@@ -8,14 +8,14 @@
 
 #include "engine/App.h"
 #include "engine/core/debug/Log.h"
-#include "engine/physics/Hitbox.h"
+#include "game/components/shooting/Hitbox.h"
 #include "engine/render/Camera.h"
 #include "engine/render/MeshRenderer.h"
 #include "engine/scene/OnUpdateEvent.h"
 #include "game/components/Controllers/AIController.h"
 #include "game/components/Controllers/PlayerController.h"
 #include "game/components/FollowCamera.h"
-#include "game/components/Shooter.h"
+#include "game/components/shooting/Shooter.h"
 #include "game/components/VehicleComponent.h"
 #include "game/components/audio/AudioEmitter.h"
 #include "game/components/audio/AudioListener.h"
